fix(ci): add Default impls for HAL pin stubs, remove unnecessary casts in timers
- hal-microrusefi: Default for MicroRusEFIIgnitionPins and MicroRusEFITriggerPins
- hal-*/timer.rs: remove redundant 'as u64' cast (Instant::as_micros() is already u64)
- cargo fmt --all

+++ hal-huge/src/timer.rs
@@ -33,7 +33,7 @@ impl Default for Stm32SystemTimer {
 
 impl SystemTimer for Stm32SystemTimer {
     fn now_us(&self) -> u64 {
-        Instant::now().as_micros() as u64
+        Instant::now().as_micros()
     }
 
     fn schedule_us(&mut self, delay_us: u64, callback: TimerCallback) {
